+++ client/mocks/fileMock.js
