+++ client/mocks/styleMock.js
