+++ client/src/containers/LoginContainer/LoginContainer.tsx
@@ -2,29 +2,30 @@ import React from "react";
 import {Link, useHistory} from "react-router-dom";
 
 import {storeToken} from "../../api/client";
-import NamePasswordForm from "../../components/NamePasswordFrom";
-import {useRegisterMutation} from "../../queries/hooks";
+import NamePasswordForm from "../../components/NamePasswordForm";
+import {useLoginMutation} from "../../queries/hooks";
 
-const RegisterContainer: React.FunctionComponent = (): JSX.Element => {
+const LoginContainer: React.FunctionComponent = (): JSX.Element => {
     const history = useHistory();
-    const [register] = useRegisterMutation({
+    const [register] = useLoginMutation({
         onCompleted: async (data) => {
-            await storeToken(data.register.token);
+            await storeToken(data.login.token);
             history.push("/talks");
         }
     });
+
     const handleSubmit = async (name: string, password: string) => {
         await register({variables: {input: {name, password}}});
     };
 
     return (
         <NamePasswordForm
-            title={"Create Account"}
-            action="register"
+            title={"Login"}
+            action={"login"}
             onSubmit={handleSubmit}
-            alternative={<Link to="/login">login</Link>}
+            alternative={<Link to="/register">don&#39;t have account?</Link>}
         />
     );
 };
 
-export default RegisterContainer;
+export default LoginContainer;
