+++ client/src/containers/TalksContainer/TalkFragment.gql
